refactor: 예약 조회 로직 서비스 분리
